% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{default_route_colors}
\alias{default_route_colors}
\title{Default colors for routes when operator color is not available}
\format{
An object of class \code{character} of length 10.
}
\usage{
default_route_colors
}
\description{
Default colors for routes when operator color is not available
}
\keyword{internal}
