% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{find_sections_near_station}
\alias{find_sections_near_station}
\title{Find sections near a station}
\usage{
find_sections_near_station(line_railroad, station_geom, max_dist_m = 500)
}
\arguments{
\item{line_railroad}{sf object with railway sections}

\item{station_geom}{sf geometry of the station}

\item{max_dist_m}{Maximum distance in meters}
}
\value{
Vector of section indices
}
\description{
Find sections near a station
}
\keyword{internal}
