% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get the cache directory for railwayJP data}
\usage{
get_cache_dir()
}
\value{
Character string with the path to the cache directory
}
\description{
Get the cache directory for railwayJP data
}
\examples{
get_cache_dir()
}
