% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Get the path to the railway data directory}
\usage{
get_data_dir(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Optional custom cache directory path}
}
\value{
Character string with the path to the data directory
}
\description{
Get the path to the railway data directory
}
