% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_info}
\alias{get_data_info}
\title{Get data info}
\usage{
get_data_info()
}
\value{
A list with data statistics
}
\description{
Returns information about the loaded railway data.
}
\examples{
\dontrun{
get_data_info()
}
}
