% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_default_color}
\alias{get_default_color}
\title{Get a default color by index}
\usage{
get_default_color(index)
}
\arguments{
\item{index}{Color index (will be recycled if > 10)}
}
\value{
Hex color code
}
\description{
Get a default color by index
}
\keyword{internal}
