% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_japan_map}
\alias{get_japan_map}
\title{Get Japan map for plotting}
\usage{
get_japan_map()
}
\value{
sf object with Japan prefecture boundaries
}
\description{
Get Japan map for plotting
}
\keyword{internal}
