% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_railroad_data}
\alias{get_railroad_data}
\title{Get railroad data from package cache}
\usage{
get_railroad_data()
}
\value{
sf object with railroad sections
}
\description{
Get railroad data from package cache
}
\keyword{internal}
