% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_station_data}
\alias{get_station_data}
\title{Get station data from package cache}
\usage{
get_station_data()
}
\value{
sf object with stations
}
\description{
Get station data from package cache
}
\keyword{internal}
