% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_stations_along_path}
\alias{get_stations_along_path}
\title{Get stations along a path of sections}
\usage{
get_stations_along_path(
  line_railroad,
  line_stations,
  section_indices,
  max_dist_m = 300
)
}
\arguments{
\item{line_railroad}{sf object with railway sections}

\item{line_stations}{sf object with stations}

\item{section_indices}{Indices of sections in the path}

\item{max_dist_m}{Maximum distance to consider a station on the path}
}
\value{
sf object with stations along the path
}
\description{
Get stations along a path of sections
}
\keyword{internal}
