% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{list_operator_colors}
\alias{list_operator_colors}
\title{List all available operator colors}
\usage{
list_operator_colors()
}
\value{
A data frame with operator names and their colors
}
\description{
List all available operator colors
}
\examples{
list_operator_colors()
}
