% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_operators}
\alias{list_operators}
\title{List all operators}
\usage{
list_operators()
}
\value{
A data frame with operator names and number of lines/stations
}
\description{
List all operators
}
\examples{
\dontrun{
list_operators()
}
}
