% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{snap_coord}
\alias{snap_coord}
\title{Round coordinates to create snap grid (for matching endpoints)}
\usage{
snap_coord(x, precision = 5)
}
\arguments{
\item{x}{Coordinate value}

\item{precision}{Number of decimal places}
}
\value{
Rounded coordinate
}
\description{
Round coordinates to create snap grid (for matching endpoints)
}
\keyword{internal}
